requirements.txt
streamlit
pandas
numpy
pypdf
sentence-transformers
faiss-cpu
openai
python-dotenv
tqdm
